Using asinf in trig_fast

+++ coretech/common/robot/src/trig_fast.cpp
@@ -12,12 +12,11 @@
 *
 *
 **/
+#include "anki/common/robot/config.h"
 #include "anki/common/types.h"
 #include "anki/common/constantsAndMacros.h"
 #include "anki/common/robot/trig_fast.h"
 
-#include <assert.h>
-#include <math.h>
 
 // For larger input values to atan, use approximations
 // at fixed steps. (Essentially extends the LUT with courser
@@ -273,7 +272,7 @@ float atan2_acc(float y, float x)
   assert(y != 0 || x != 0);
   
   float arg = y/x;
-  float atan_val = asin( arg / sqrt(arg*arg + 1));
+  float atan_val = asinf( arg / sqrtf(arg*arg + 1));
   
   if (x > 0) {
     return atan_val;
